+++ tvos-darkmode/SceneDelegate.swift
@@ -1,0 +1,6 @@
+import UIKit
+
+@available(tvOS 13.0, *)
+final class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+    var window: UIWindow?
+}